Enhanced context switch system. Added `int 0x82` for returning from process end. Replaced CURRENT_PROCESS linked list with PROCESS_LIST vector that acts as a stack.

+++ run.sh
@@ -4,6 +4,8 @@ FILESYSTEM_HEAD="build/head.bin"
 FILESYSTEM="build/filesystem.bin"
 OS_FILE="build/os.img"
 
+die() { echo "$*" 1>&2 ; exit 1; }
+
 filesize_in_sectors() {
     SIZE_IN_BYTES=0
     if [[ "$OSTYPE" == "linux-gnu" ]]; then
@@ -39,21 +41,28 @@ create_filesystem() {
 
     # Generating binary files and adding them to `build/isofiles`` folder
     make filesystem
+    success=$?
 
-    # Mounting the file and copying files to it
-    if [[ "$OSTYPE" == "linux-gnu" ]]; then
-        sudo mount -o loop $FILESYSTEM /mnt
-        sudo cp -r build/isofiles/. /mnt
-    elif [[ "$OSTYPE" == "darwin"* ]]; then
-        MOUNT_POINT=$(hdiutil attach -imagekey diskimage-class=CRawDiskImage -nomount $FILESYSTEM)
-        echo "Mounting $FILESYSTEM to $MOUNT_POINT"
-	    sudo mount -t msdos $MOUNT_POINT /mnt
-	    sudo cp -r build/isofiles/. /mnt
-	    hdiutil detach $MOUNT_POINT > /dev/null
+    if [ $success -ne 2 ]; then
+        # Mounting the file and copying files to it
+        if [[ "$OSTYPE" == "linux-gnu" ]]; then
+            sudo mount -o loop $FILESYSTEM /mnt
+            sudo cp -r build/isofiles/. /mnt
+        elif [[ "$OSTYPE" == "darwin"* ]]; then
+            MOUNT_POINT=$(hdiutil attach -imagekey diskimage-class=CRawDiskImage -nomount $FILESYSTEM)
+            echo "Mounting $FILESYSTEM to $MOUNT_POINT"
+            sudo mount -t msdos $MOUNT_POINT /mnt
+            sudo cp -r build/isofiles/. /mnt
+            hdiutil detach $MOUNT_POINT > /dev/null
+        fi
+        # Unmounting
+        (sudo umount /mnt || true) 2> /dev/null
+        rm -r build/isofiles
+
+        return 1
+    else
+        return 2
     fi
-    # Unmounting
-    (sudo umount /mnt || true) 2> /dev/null
-    rm -r build/isofiles
 }
 
 run() {
@@ -66,15 +75,20 @@ run() {
         BLOCK_SIZE="1m"
         LD="i686-elf-ld"
     fi
+
     echo "Compiling Kernel and Bootloader..."
     # Compiling bootloader and kernel into a single file
     make head LD=$LD
     success=$?
+    
+    echo "Creating filesystem."
+    # Creating a filesystem Fat32 file with reserved sectors the size of the bootloader+kernel
+    create_filesystem
+    success2=$?
+
     # If there is no error in compiling the kernel
-    if [ $success -ne 2 ]; then
-        echo "Creating filesystem."
-        # Creating a filesystem Fat32 file with reserved sectors the size of the bootloader+kernel
-        create_filesystem
+    if [ $success -ne 2 ] && [ $success2 -ne 2 ]; then
+        echo "Compiled Kernel and Bootloader..."
         echo "Created filesystem."
         echo "Attaching kernel."
         # Attaching the kernel to the filesystem file
